Barchart calorie gram

+++ backend/src/main/java/org/codecool/fitnesstracker/fitnesstracker/controller/dto/CalorieForAnalyticsDTO.java
@@ -3,5 +3,5 @@ package org.codecool.fitnesstracker.fitnesstracker.controller.dto;
 import java.time.LocalDateTime;
 import java.util.List;
 
-public record CalorieForAnalyticsDTO(int calorie, LocalDateTime consumptionTime) {
+public record CalorieForAnalyticsDTO(double calorie, LocalDateTime consumptionTime) {
 }
